refactor(diff): move worker-shared types and helpers to highlight-worker-shared
- Extracted djb2 and highlight-worker types from highlight-worker.ts into a new highlight-worker-shared.ts,
  ensuring the main thread never imports from the worker chunk and that hash algorithms/types never drift.
- Updated DiffSurface.tsx and use-worker-highlight.ts to import from the new shared module instead of the worker directly.
- Added robust crash/message error handling in use-worker-highlight.ts to catch unavailable worker scenarios and fail open,
  preventing deadlocks on Shiki highlighting for large diffs.
- Removed redundant max line cap constant from highlight-worker.ts, relying on the shared config within shikiDiffHighlighter.
